# openapi.yaml
openapi: 3.1.0
info:
  title: MedicalTime API
  version: 1.0.0
  description: API para lembretes de medicação, eventos de tomada, dispositivos, S.O.S. e autenticação.
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Reminder:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        purpose: { type: string, nullable: true }
        description: { type: string, nullable: true }
        pricePaid:
          { type: string, nullable: true, description: "Decimal como string" }
        photoUrl: { type: string, nullable: true }
        isActive: { type: boolean }
        schedules:
          type: array
          items: { $ref: "#/components/schemas/Schedule" }
      required: [id, name, isActive]
    ReminderCreate:
      type: object
      required: [name, schedules]
      properties:
        name: { type: string }
        purpose: { type: string }
        description: { type: string }
        pricePaid: { type: string, description: "Decimal como string" }
        photoUrl: { type: string, format: uri }
        schedules:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/ScheduleCreate" }
    ReminderUpdate:
      type: object
      properties:
        name: { type: string }
        purpose: { type: string }
        description: { type: string }
        pricePaid: { type: string }
        photoUrl: { type: string, format: uri }
        isActive: { type: boolean }
    Schedule:
      type: object
      properties:
        id: { type: string }
        ingestionTimeMinutes: { type: integer, minimum: 0, maximum: 1439 }
        daysOfWeekBitmask:
          {
            type: integer,
            minimum: 0,
            maximum: 127,
            description: "1=Dom,2=Seg,4=Ter,8=Qua,16=Qui,32=Sex,64=Sáb; 0=todos",
          }
        isActive: { type: boolean }
      required: [id, ingestionTimeMinutes, daysOfWeekBitmask, isActive]
    ScheduleCreate:
      type: object
      required: [ingestionTimeMinutes]
      properties:
        ingestionTimeMinutes: { type: integer, minimum: 0, maximum: 1439 }
        daysOfWeekBitmask: { type: integer, minimum: 0, maximum: 127 }
        isActive: { type: boolean }
    ScheduleUpdate:
      type: object
      properties:
        ingestionTimeMinutes: { type: integer, minimum: 0, maximum: 1439 }
        daysOfWeekBitmask: { type: integer, minimum: 0, maximum: 127 }
        isActive: { type: boolean }
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object, nullable: true }
          required: [code, message]
      required: [error]
  responses:
...
    NotFoundError:
      description: Não encontrado
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          examples:
            generico:
              summary: Recurso inexistente
              value:
                error:
                  code: NOT_FOUND
                  message: "Recurso não encontrado"
...
    BadRequestError:
      description: Requisição inválida
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          examples:
            validacao:
              summary: Falha de validação de campos
              value:
                error:
                  code: BAD_REQUEST
                  message: "Falha de validação"
                  details:
                    issues:
                      - path: ["email"]
                        message: "E-mail inválido"
                      - path: ["password"]
                        message: "Mínimo de 6 caracteres"
...
    InternalServerError:
      description: Erro interno
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          examples:
            generico:
              summary: Erro inesperado
              value:
                error:
                  code: INTERNAL_ERROR
                  message: "Erro interno inesperado"
security:
  - bearerAuth: []
paths:
...
  /reminders:
    get:
      operationId: listReminders
      summary: Listar lembretes (array simples)
      responses:
        "200":
          description: Lista de lembretes
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Reminder" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    post:
      operationId: createReminder
      summary: Criar lembrete
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReminderCreate" }
            examples:
              basico:
                summary: Lembrete diário simples
                value:
                  name: "Vitamina D"
                  purpose: "Saúde óssea"
                  description: "Cápsula de 2000 UI"
                  schedules:
                    - ingestionTimeMinutes: 480
                      daysOfWeekBitmask: 0
      responses:
        "201": { description: Criado }
        "400": { $ref: "#/components/responses/BadRequestError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /reminders/{id}:
    get:
      operationId: getReminder
      summary: Obter lembrete
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Ok }
        "400": { $ref: "#/components/responses/BadRequestError" }
        "404": { $ref: "#/components/responses/NotFoundError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    patch:
      operationId: updateReminder
      summary: Atualizar lembrete
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReminderUpdate" }
      responses:
        "200": { description: Ok }
        "400": { $ref: "#/components/responses/BadRequestError" }
        "404": { $ref: "#/components/responses/NotFoundError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    delete:
      operationId: deleteReminder
      summary: Desativar (soft delete) lembrete
      description: Marca o lembrete e seus schedules como inativos (isActive=false). Idempotente.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "204": { description: Desativado }
        "400": { $ref: "#/components/responses/BadRequestError" }
        "404": { $ref: "#/components/responses/NotFoundError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /reminders/{id}/schedules:
    post:
      operationId: addReminderSchedule
      summary: Adicionar schedule a um lembrete
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ScheduleCreate" }
            examples:
              manha_segqua:
                summary: Dose às 08:30 Seg e Qua
                value:
                  ingestionTimeMinutes: 510
                  daysOfWeekBitmask: 2 + 8
      responses:
        "201": { description: Criado }
        "400": { $ref: "#/components/responses/BadRequestError" }
        "404": { $ref: "#/components/responses/NotFoundError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /reminders/schedules/{scheduleId}:
    patch:
      operationId: updateSchedule
      summary: Atualizar schedule
      parameters:
        - in: path
          name: scheduleId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ScheduleUpdate" }
      responses:
        "200": { description: Ok }
        "400": { $ref: "#/components/responses/BadRequestError" }
        "404": { $ref: "#/components/responses/NotFoundError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    delete:
      operationId: deleteSchedule
      summary: Remover schedule
      parameters:
        - in: path
          name: scheduleId
          required: true
          schema: { type: string }
      responses:
        "204": { description: Removido }
        "400": { $ref: "#/components/responses/BadRequestError" }
        "404": { $ref: "#/components/responses/NotFoundError" }
        "500": { $ref: "#/components/responses/InternalServerError" }
...